+++ creditcards/corporate-creditcard/constant.js
@@ -2,21 +2,6 @@
 // impoted as CC_CONSTANT key name in all files.
 
 const JOURNEY_NAME = 'CORPORATE_CARD_JOURNEY';
-const ID_COM = {
-  productCode: 'CORPCC',
-  scopeMap: {
-    only_casa: {
-      no: 'AACC',
-      yes: 'ADOBE_PACC',
-    },
-    casa_asset: {
-      no: 'AACC',
-      yes: 'ADOBE_PACC',
-    },
-    casa_cc: 'PADC',
-    only_cc: 'OYCC',
-  },
-};
 
 const DOM_ELEMENT = {
   identifyYourself: {
@@ -60,6 +45,5 @@ const DOM_ELEMENT = {
 };
 export {
   JOURNEY_NAME,
-  ID_COM,
   DOM_ELEMENT,
 };
